Migrate over to using beacon_person table for now until we have auth in place
