--- src/main/java/uk/gov/mca/beacons/service/model/Person.java
@@ -1,137 +1,0 @@
-package uk.gov.mca.beacons.service.model;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-/**
- * 1. create a new person object per beacon
- * 2. lookup is there already a user with (insert criteria, i.e: same email)?
- */
-@Entity
-@EntityListeners(AuditingEntityListener.class)
-public class Person {
-
-  @Id
-  @GeneratedValue
-  private UUID id;
-
-  private String name;
-
-  private String telephone;
-
-  private String email;
-
-  @CreatedDate
-  private LocalDateTime createdDate;
-
-  @Column(name = "address_line_1")
-  private String addressLine1;
-
-  @Column(name = "address_line_2")
-  private String addressLine2;
-
-  @Column(name = "address_line_3")
-  private String addressLine3;
-
-  @Column(name = "address_line_4")
-  private String addressLine4;
-
-  private String postcode;
-
-  private String county;
-
-  public UUID getId() {
-    return id;
-  }
-
-  public void setId(UUID id) {
-    this.id = id;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public String getTelephone() {
-    return telephone;
-  }
-
-  public void setTelephone(String telephone) {
-    this.telephone = telephone;
-  }
-
-  public String getEmail() {
-    return email;
-  }
-
-  public void setEmail(String emailAddress) {
-    this.email = emailAddress;
-  }
-
-  public LocalDateTime getCreatedDate() {
-    return createdDate;
-  }
-
-  public void setCreatedDate(LocalDateTime createdDate) {
-    this.createdDate = createdDate;
-  }
-
-  public String getAddressLine1() {
-    return addressLine1;
-  }
-
-  public void setAddressLine1(String addressLine1) {
-    this.addressLine1 = addressLine1;
-  }
-
-  public String getAddressLine2() {
-    return addressLine2;
-  }
-
-  public void setAddressLine2(String addressLine2) {
-    this.addressLine2 = addressLine2;
-  }
-
-  public String getAddressLine3() {
-    return addressLine3;
-  }
-
-  public void setAddressLine3(String addressLine3) {
-    this.addressLine3 = addressLine3;
-  }
-
-  public String getAddressLine4() {
-    return addressLine4;
-  }
-
-  public void setAddressLine4(String addressLine4) {
-    this.addressLine4 = addressLine4;
-  }
-
-  public String getPostcode() {
-    return postcode;
-  }
-
-  public void setPostcode(String postcode) {
-    this.postcode = postcode;
-  }
-
-  public String getCounty() {
-    return county;
-  }
-
-  public void setCounty(String country) {
-    this.county = country;
-  }
-}
